:+1: Manually define `nvim_win_get_config` in Neovim
