+++ function/nvim/_manual.ts
@@ -1,1 +1,2 @@
 export * from "./nvim_open_win.ts";
+export * from "./nvim_win_get_config.ts";
